third_party/WebKit/LayoutTests/external/wpt/html/rendering/non-replaced-elements/flow-content-0/form-margin-quirk-expected.txt

This is a testharness.js-based test.
FAIL form margin quirk assert_equals: marginRight expected "16px" but got "0px"
Harness: the test ran to completion.
